All non-MKL parts are now passing the integration tests after the merge. Trying again
Former-commit-id: 6f552ea754022cbf513f02a9002f27a1d62db1d4

+++ gadgets/grappa/GrappaCalibrationBuffer.h
@@ -124,7 +124,7 @@ class EXPORTGADGETSGRAPPA GrappaCalibrationBuffer
 {
 
  public:
-  GrappaCalibrationBuffer(std::vector<unsigned int> dimensions, 
+  GrappaCalibrationBuffer(std::vector<size_t> dimensions, 
 			  boost::shared_ptr< GrappaWeights<float> > w,
 			  GrappaWeightsCalculator<float>* weights_calculator);
   virtual ~GrappaCalibrationBuffer() {}
@@ -133,7 +133,7 @@ class EXPORTGADGETSGRAPPA GrappaCalibrationBuffer
 
  private:
   hoNDArray< std::complex<float> > buffer_;
-  std::vector<unsigned int> dimensions_;
+  std::vector<size_t> dimensions_;
   boost::shared_ptr< GrappaWeights<float> > weights_;
   GrappaWeightsCalculator<float>* weights_calculator_;
   CalibrationBufferCounter buffer_counter_;
