+++ toolboxes/core/gpu/hoCuNDArray_blas.h
@@ -1,4 +1,5 @@
 #pragma once
+
 #include "hoCuNDArray.h"
 #include "cudaDeviceManager.h"
 #include "gpucore_export.h"
@@ -17,7 +18,7 @@ namespace Gadgetron{
    * @param x Input data
    * @return index of absolute minimum values
    */
-  template<class T> EXPORTGPUCORE int amin( hoCuNDArray<T> *x);
+  template<class T> EXPORTGPUCORE size_t amin( hoCuNDArray<T> *x);
   
   /**
    * @brief Gets the index of the index of the element with maximum absolute
@@ -25,7 +26,7 @@ namespace Gadgetron{
    * @return index of absolute maximum values
    * @details Note that this returns the C-style index and NOT the Fortran index.
    */
-  template<class T> EXPORTGPUCORE int amax( hoCuNDArray<T> *x );
+  template<class T> EXPORTGPUCORE size_t amax( hoCuNDArray<T> *x );
   
   template<class T> EXPORTGPUCORE typename realType<T>::Type asum( hoCuNDArray<T> *x );
 }
